# swagger.yml
swagger: '2.0'
info:
  title: Movie and User API
  description: API for managing movies and users
  version: '1.0'
basePath: /
schemes:
  - http
  - https
tags:
  - name: Movies
    description: Operations related to movies
  - name: Users
    description: Operations related to users
paths:
  /movies:
    get:
      tags:
        - Movies
      summary: Get a list of movies with pagination
      parameters:
        - name: page
          in: query
          type: integer
          required: true
          description: Page number
        - name: limit
          in: query
          type: integer
          required: true
          description: Items per page
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/Movie'
        '500':
          description: Database error
    post:
      tags:
        - Movies
      summary: Create a new movie
      parameters:
        - name: Movie
          in: body
          required: true
          schema:
            $ref: '#/definitions/Movie'
      responses:
        '201':
          description: Movie created successfully
        '500':
          description: Database error
  /movies/{id}:
    get:
      tags:
        - Movies
      summary: Get a movie by ID
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: Movie ID
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/Movie'
        '404':
          description: Movie not found
        '500':
          description: Database error
    put:
      tags:
        - Movies
      summary: Update a movie by ID
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: Movie ID
        - name: Movie
          in: body
          required: true
          schema:
            $ref: '#/definitions/Movie'
      responses:
        '200':
          description: Movie updated successfully
        '404':
          description: Movie not found
        '500':
          description: Database error
    delete:
      tags:
        - Movies
      summary: Delete a movie by ID
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: Movie ID
      responses:
        '200':
          description: Movie deleted successfully
        '404':
          description: Movie not found
        '500':
          description: Database error
  /users/register:
    post:
      tags:
        - Users
      summary: Register a new user
      parameters:
        - name: User
          in: body
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Email already exists
        '500':
          description: Database error
  /users:
    get:
      tags:
        - Users
      summary: Get a list of users with pagination
      parameters:
        - name: page
          in: query
          type: integer
          required: true
          description: Page number
        - name: limit
          in: query
          type: integer
          required: true
          description: Items per page
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        '500':
          description: Database error
  /users/login:
    post:
      tags:
        - Users
      summary: User login to obtain a token
      parameters:
        - name: User Login
          in: body
          required: true
          schema:
            $ref: '#/definitions/UserLogin'
      responses:
        '200':
          description: User logged in successfully
          schema:
            type: object
            properties:
              token:
                type: string
        '400':
          description: Invalid credentials
        '500':
          description: Database error
definitions:
  Movie:
    type: object
    properties:
      id:
        type: integer
      title:
        type: string
      genres:
        type: string
      year:
        type: integer
  UserRegister:
    type: object
    properties:
      id:
        type: integer
      email:
        type: string
      password:
        type: string
      role:
        type: string
  UserLogin:
    type: object
    properties:
      email:
        type: string
      password:
        type: string